% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abunMatrix}
\alias{abunMatrix}
\title{Larger Sample Size Expression Abundance Matrix From C.elegans Quantification Result}
\format{
A data frame with 21390 rows and 18 variables.
\describe{
\item{cg_x/sg_x}{gain of function sequences}
\item{cl_x/sl_x}{loss of function sequences}
\item{cw_x/sw_x}{wild type sequences}
}
}
\source{
Lunenfeld Tanenbaum Research Institute
}
\usage{
abunMatrix
}
\description{
An RNAseq experiment conducted using C.elegans from 2021-2022 in Canada.
18 samples are included.
}
\examples{
\dontrun{
abunMatrix
}
}
\keyword{datasets}
