% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditionsDF}
\alias{conditionsDF}
\title{Genotypes for C.elegans samples}
\format{
A data frame with 3 rows and 1 variable:
\describe{
\item{WT_WC_1}{ild type C.elegans sample sequence}
\item{lf_WC_1}{loss of function C.elegans sample sequence}
\item{gf_WC_1}{gain of function C.elegans sample sequence}
\item{genotype}{indicating the genotype of the sample.}
}
\source{
Lunenfeld Tanenbaum Research Institute
}
\usage{
conditionsDF
}
\description{
An example file indicating the genotype conditions of C.elegans samples. Used
for displaying \code{plotPCA()} function utility.
}
\keyword{datasets}
