% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countMatrix}
\alias{countMatrix}
\title{Expression Abundance Matrix From C.elegans Quantification Result}
\format{
A data frame with 21390 rows and 3 variables:
\describe{
\item{WT_WC_1}{ild type C.elegans sample sequence}
\item{lf_WC_1}{loss of function C.elegans sample sequence}
\item{gf_WC_1}{gain of function C.elegans sample sequence}
}
\source{
Lunenfeld Tanenbaum Research Institute
}
\usage{
countMatrix
}
\description{
An RNAseq experiment conducted using C.elegans from 2021-2022 in Canada.
}
\keyword{datasets}
