% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Download salmon software through bioconda.}
\usage{
install()
}
\description{
A function that installs Salmon software for indexing and quantification.
If conda is not previously installed, the function would install for use of
Salmon.
}
\references{
Ushey K, Allaire J, Wickham H, Ritchie G (2022). rstudioapi: Safely Access
the RStudio API.
\href{https://rstudio.github.io/rstudioapi/}
\href{https://github.com/rstudio/rstudioapi}
}
