% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installSalmon.R
\name{installSalmon}
\alias{installSalmon}
\title{Download salmon software through bioconda.}
\usage{
installSalmon()
}
\description{
A function that installs Salmon software for indexing and quantification.
If conda is not previously installed, the function would install for use of
Salmon.
}
\examples{
\dontrun{
installSalmon()
}

}
\references{
Ushey K, Allaire J, Wickham H, Ritchie G (2022). rstudioapi: Safely Access
the RStudio API.
\href{https://rstudio.github.io/rstudioapi/}{Link}
\href{https://github.com/rstudio/rstudioapi}{Link}
}
