% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainFiles.R
\name{obtainCDNA}
\alias{obtainCDNA}
\title{Obtain complete cDNA fa.gz file for a specific species from Ensembl}
\usage{
obtainCDNA(species = NA, wantedVersion = NA, download = F)
}
\arguments{
\item{species}{A string indicating the specific name of the species.
The default value is NA.}

\item{wantedVersion}{A double indicating the version of Ensembl archive.
The default value is NA.}

\item{download}{A boolean indicating whether to download the file or not.
The default value is FALSE.}
}
\value{
Returns the name of the fa.gz compressed file of cDNA.
}
\description{
A function that extracts complete cDNA fa.gz file from Ensembl. If a release
version is provided, the data will be extracted from the given version. If
the release version is not provided, then the default version would be the
latest version of Ensembl Archive.
}
\examples{
\dontrun{
obtainCDNA(species = "Caenorhabditis Elegans", wantedVersion = 107)
}


}
\references{
Ensembl 2022, Nucleic Acids Research, Volume 50, Issue D1,
7 January 2022, Pages D988–D995,
\href{https://doi.org/10.1093/nar/gkab1049}{Link}

Wickham H (2022). rvest: Easily Harvest (Scrape) Web Pages.
\href{ https://CRAN.R-project.org/package=rvest}{Link}

Wickham H (2022). stringr: Simple, Consistent Wrappers for
Common String Operations.
\href{https://CRAN.R-project.org/package=stringr}{Link}
}
