% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainDNA.R
\name{obtainDNA}
\alias{obtainDNA}
\title{Obtain DNA toplevel fa.gz file for a specific species from Ensembl}
\usage{
obtainDNA(species, wantedVersion = NA, download = F)
}
\arguments{
\item{species}{A string indicating the specific name of the species.}

\item{wantedVersion}{A double indicating the version of Ensembl archive.}

\item{download}{A boolean indicating whether to download the file or not.}
}
\value{
Returns the name of the fa.gz compressed file of DNA.
}
\description{
A function that extracts DNA toplevel fa.gz file from Ensembl. If a release
version is provided, the data will be extracted from the given version. If
the release version is not provided, then the default version would be the
latest version of Ensembl Archive.
}
\references{
Ensembl 2022, Nucleic Acids Research, Volume 50, Issue D1,
7 January 2022, Pages D988–D995,
\href{https://doi.org/10.1093/nar/gkab1049}{Link}

Wickham H (2022). rvest: Easily Harvest (Scrape) Web Pages.
\href{https://rvest.tidyverse.org/}
\href{https://github.com/tidyverse/rvest}

Wickham H (2022). stringr: Simple, Consistent Wrappers for
Common String Operations.
\href{http://stringr.tidyverse.org}
\href{https://github.com/tidyverse/stringr}
}
