% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatrix.R
\name{obtainGTF}
\alias{obtainGTF}
\title{Obtain GTF file of a specific species from Ensmbl latest version}
\usage{
obtainGTF(species, wantedVersion = NA)
}
\arguments{
\item{species}{A string indicating the specific name of the species to get
reference transcriptome.}

\item{wantedVersion}{A double indicating the version of Ensembl archive.}
}
\value{
Returns the name of the gtf.gz compressed file of the
reference transcriptome.
}
\description{
Helper function for generateMatrix to obtain reference transcriptome from
the latest version of Ensembl.
}
