% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMatrix.R
\name{plotHeatMap}
\alias{plotHeatMap}
\title{Plot abundance heatmap}
\usage{
plotHeatMap(matrix = NULL, head = T)
}
\arguments{
\item{matrix}{A count matrix. Default is NULL.}

\item{head}{A boolean indicating whether to plot only the first 6 genes in
the count matrix, or plot all.}
}
\value{
A heatmap of the count matrix.
}
\description{
Plot a heatmap to visualize the transcript abundance among wild type and
mutant samples.
}
\examples{
plotHeatMap(AbunRNA::abunMatrix)

}
\references{
Raivo Kolde \href{https://CRAN.R-project.org/package=pheatmap}{Link}
}
