% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Principle component analysis on input matrix}
\usage{
plotPCA(mat, scale = TRUE, conditions, col, x = 1, y = 2)
}
\arguments{
\item{mat}{A count matrix}

\item{scale}{A boolean indicating whether to scale the variables (divide by
standard deviation). Default value is set to TRUE}

\item{conditions}{A data frame indicating the conditions of the sample}

\item{col}{A string indicating the category where observations should be
grouped into. This should be a column name in "conditions" data frame.}

\item{x}{An integer indicating the x variable (e.g. x = 1 would extrace PC1
from principle component table)}

\item{y}{An integer indicating the y variable}
}
\value{
Returns a list containing the principle component analysis results,
and a PCA plot for x and y variables.
}
\description{
A function that performs principle component analysis on input matrix, and
plot the graph indicaiting the results, grouped in category given by
the user.
}
\examples{
graphPlot <- plotPCA(mat = countMatrix, scale = TRUE, conditions = conditionsDF,
                     col = "genotype")
graphPlot$PCA
graphPlot$Plot

}
\references{
Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4
\href{https://ggplot2.tidyverse.org}

R Core Team (2013). R: A language and environment for statistical computing.
R Foundation for Statistical Computing, Vienna, Austria. ISBN 3-900051-07-0
\href{http://www.R-project.org/}
}
