% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbObj.R
\name{txdbObj}
\alias{txdbObj}
\title{Generate txdb object.}
\usage{
txdbObj(sfSeq = NA, refTrp = NA, species = NA, release = NA, key = "TXNAME")
}
\arguments{
\item{sfSeq}{A character vector indicating the names of salmon
output .sf files. The default value is NA.}

\item{refTrp}{A string indicating the path to the reference
transcriptome file. The default value is NA.}

\item{species}{A string indicating the specific name of the species to get
reference transcriptome. The default value is NA.}

\item{release}{A string indicating the release version of the reference.
Default is the current release. The default value is NA.}

\item{key}{A stirng indicating the keytype to be used. The default value
is "TXNAME".}
}
\value{
Returns a txdb object
}
\description{
A function that generates a txdb object from input quantification files,
which can later be used for generating count matrix of transcript abundance
and the downstream expression analysis pipeline.
}
\examples{
cw1Quants <- system.file("extdata",
                          "cw1_quants",
                          "quant.sf",
                          package = "AbunRNA")
cl1Quants <- system.file("extdata",
                          "cl1_quants",
                          "quant.sf",
                          package = "AbunRNA")
cg1Quants <- system.file("extdata",
                          "cg1_quants",
                          "quant.sf",
                          package = "AbunRNA")
sfSe <- c(cw1Quants, cl1Quants, cg1Quants)
# Example 1:
# Directly obtain the latest reference transcriptome from Ensembl database
\dontrun{
txi <- txdbObj(sfSeq = sfSe,
                species = "Caenorhabditis elegans",
                key = "TXNAME")
}

# Example 2:
# Directly obtain version 107 reference transcriptome from Ensembl database
\dontrun{
txi <- txdbObj(sfSeq = sfSe,
                species = "Caenorhabditis elegans",
                release = 107,
                key = "TXNAME")
}

}
\references{
Ensembl 2022, Nucleic Acids Research, Volume 50, Issue D1,
7 January 2022, Pages D988–D995,
\href{https://doi.org/10.1093/nar/gkab1049}{Link}

Wickham H (2022). rvest: Easily Harvest (Scrape) Web Pages.
\href{https://rvest.tidyverse.org/}{Link}
\href{https://github.com/tidyverse/rvest}{Link}

Durinck S, Spellman P, Birney E, Huber W (2009). “Mapping identifiers for the
integration of genomic datasets with the R/Bioconductor package biomaRt.”
Nature Protocols, 4, 1184–1191.

Durinck S, Moreau Y, Kasprzyk A, Davis S, De Moor B, Brazma A, Huber W
(2005). “BioMart and Bioconductor: a powerful link between biological
databases and microarray data analysis.” Bioinformatics, 21, 3439–3440.

Lawrence M, Huber W, Pagès H, Aboyoun P, Carlson M, Gentleman R, Morgan M,
Carey V (2013). “Software for Computing and Annotating Genomic Ranges.” PLoS
Computational Biology, 9. doi: 10.1371/journal.pcbi.1003118,
}
