% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abunMatrix}
\alias{abunMatrix}
\title{Larger Sample Size Expression Abundance Matrix From C.elegans Quantification Result}
\format{
A data frame with 21390 rows and 18 variables.
\describe{
\item{cg1}{gain of function sequence sample}
\item{cg2}{gain of function sequence sample}
\item{cg3}{gain of function sequence sample}
\item{sg1}{gain of function sequence sample}
\item{sg2}{gain of function sequence sample}
\item{sg3}{gain of function sequence sample}
\item{cl1}{loss of function sequence sample}
\item{cl2}{loss of function sequence sample}
\item{cl3}{loss of function sequence sample}
\item{sl1}{loss of function sequence sample}
\item{sl2}{loss of function sequence sample}
\item{sl3}{loss of function sequence sample}
\item{cw1}{wild type sequence sample}
\item{cw2}{wild type sequence sample}
\item{cw3}{wild type sequence sample}
\item{sw1}{wild type sequence sample}
\item{sw2}{wild type sequence sample}
\item{sw3}{wild type sequence sample}
}
}
\source{
Lunenfeld Tanenbaum Research Institute
}
\usage{
abunMatrix
}
\description{
An RNAseq experiment conducted using C.elegans from 2021-2022 in Canada.
18 samples are included.
}
\examples{
\dontrun{
abunMatrix
}
}
\keyword{datasets}
