% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abunMatrix}
\alias{abunMatrix}
\title{Larger Sample Size Expression Abundance Matrix From C.elegans Quantification Result}
\format{
A data frame with 21390 rows and 18 variables.
\describe{
\item{cg_1}{gain of function sequence sample}
\item{cg_2}{gain of function sequence sample}
\item{cg_3}{gain of function sequence sample}
\item{sg_1}{gain of function sequence sample}
\item{sg_2}{gain of function sequence sample}
\item{sg_3}{gain of function sequence sample}
\item{cl_1}{loss of function sequence sample}
\item{cl_2}{loss of function sequence sample}
\item{cl_3}{loss of function sequence sample}
\item{sl_1}{loss of function sequence sample}
\item{sl_2}{loss of function sequence sample}
\item{sl_3}{loss of function sequence sample}
\item{cw_1}{wild type sequence sample}
\item{cw_2}{wild type sequence sample}
\item{cw_3}{wild type sequence sample}
\item{sw_1}{wild type sequence sample}
\item{sw_2}{wild type sequence sample}
\item{sw_3}{wild type sequence sample}
}
}
\source{
Lunenfeld Tanenbaum Research Institute
}
\usage{
abunMatrix
}
\description{
An RNAseq experiment conducted using C.elegans from 2021-2022 in Canada.
18 samples are included.
}
\examples{
\dontrun{
abunMatrix
}
}
\keyword{datasets}
