% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigCond}
\alias{bigCond}
\title{Genotypes for C.elegans larger samples}
\format{
A data frame with 18 rows and 2 variables.
\describe{
\item{lf}{gain of function genotype}
\item{cl_x/sl_x}{loss of function genotype}
\item{cw_x/sw_x}{wild type genotype}
}
}
\source{
Lunenfeld Tanenbaum Research Institute
}
\usage{
bigCond
}
\description{
A condition matrix for samples in abunMatrix, describing the genotype
of each sample.
18 samples are included.
}
\examples{
\dontrun{
bigCond
}
}
\keyword{datasets}
