% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditionsDF}
\alias{conditionsDF}
\title{Genotypes for C.elegans samples}
\format{
A data frame with 3 rows and 1 variable:
\describe{
\item{genotype}{indicating the genotype of the sample i.e. wild type,
loss of function, gain of function}
\item{samples}{sample names}
}
}
\source{
Lunenfeld Tanenbaum Research Institute
}
\usage{
conditionsDF
}
\description{
An example file indicating the genotype conditions of C.elegans samples. Used
for displaying \code{plotPCA()} function utility.
}
\examples{
\dontrun{
conditionsDF
}
}
\keyword{datasets}
