% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainFiles.R
\name{obtainGTF}
\alias{obtainGTF}
\title{Obtain GTF file of the genome annotation for a specific species from Ensembl}
\usage{
obtainGTF(species = NA, wantedVersion = NA, download = F)
}
\arguments{
\item{species}{A string indicating the specific name of the species to get
reference transcriptome. The default value is NA.}

\item{wantedVersion}{A double indicating the version of Ensembl archive.
The default value is NA.}

\item{download}{A boolean indicating whether to download the file or not.
The default value is FALSE.}
}
\value{
Returns the name of the gtf.gz compressed file of the
reference transcriptome.
}
\description{
A function that extracts reference transcriptome from Ensembl. If a release
version is provided, the data will be extracted from the given version. If
the release version is not provided, then the default version would be the
latest version of Ensembl Archive.
}
\examples{
\dontrun{
obtainGTF("Caenorhabditis Elegans", wantedVersion=107)
}

}
\references{
Ensembl 2022, Nucleic Acids Research, Volume 50, Issue D1,
7 January 2022, Pages D988–D995,
\href{https://doi.org/10.1093/nar/gkab1049}{Link}

Wickham H (2022). rvest: Easily Harvest (Scrape) Web Pages.
\href{https://rvest.tidyverse.org/}{Link}
\href{https://github.com/tidyverse/rvest}{Link}

Wickham H (2022). stringr: Simple, Consistent Wrappers for
Common String Operations.
\href{http://stringr.tidyverse.org}{Link}
\href{https://github.com/tidyverse/stringr}{Link}
}
