% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{quantification}
\alias{quantification}
\title{Indexing and quantification.}
\usage{
quantification(
  species = NA,
  release = NA,
  indexName = "index",
  fastq,
  quantOut = "quantOutput"
)
}
\arguments{
\item{species}{A string indicating the specific name of the species to get
reference transcriptome.}

\item{release}{A string indicating the release version of the reference.
Default is the current release.}

\item{indexName}{A string indicating the name of the output index file.}

\item{fastq}{A character vector indicating the fastq files of sequences.}

\item{quantOut}{A character vector indicating the names for quantification files.}
}
\value{
Returns quantification files generated by Salmon program saved in
quantOut directory.
}
\description{
A function that computes indexing and quantification using default settings
of Salmon. Make sure Salmon has been installed. If not, run
\code{install()} function before calling \code{quantification()}.
}
\examples{
\dontrun{
quantification(species = "Caenorhabditis Elegans",
              release = NA,
              indexName = "celegansINDEX",
              fastq = "celegans.fastq",
              quantOut = "celegansQUANT")
}

}
\references{
Patro, R., Duggal, G., Love, M. et al. Salmon provides fast and bias-aware
quantification of transcript expression. Nat Methods 14, 417–419 (2017).
\href{https://doi.org/10.1038/nmeth.4197}{Link}

Ushey K, Allaire J, Wickham H, Ritchie G (2022). rstudioapi: Safely Access
the RStudio API.
\href{https://CRAN.R-project.org/package=rstudioapi}{Link}
}
