% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAbunRNA.R
\name{runAbunRNA}
\alias{runAbunRNA}
\title{Launch Shiny App for AbunRNA}
\usage{
runAbunRNA()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for AbunRNA.
The shiny app includes demo for displaying two count matrices, plot heatmap
and perform principle component analysis. Users can also upload their own
quantification files to generate their own matrix.
}
\examples{
\dontrun{

AbunRNA::runAbunRNA()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
